save me please

+++ handbook-dao/src/main/java/com/qmis/dao/DepartmentRepository.java
@@ -1,9 +1,9 @@
 package com.qmis.dao;
 
-import com.qmis.entyti.Department;
+import com.qmis.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartmentRepository extends JpaRepository<Department,Integer> {
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 }
