+++ handbook-dao/src/main/java/com/qmis/dao/EmployeeRepository.java
@@ -1,9 +1,9 @@
 package com.qmis.dao;
 
-import com.qmis.entyti.Employee;
+import com.qmis.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
